+++ src/astnode.cpp
@@ -1,0 +1,11 @@
+#include "pdc.h"
+
+struct ASTNode {
+	PseudoValue value;
+	void (*resolver)(Runtime* r, ASTNode* self);
+	ASTNode* left;
+	ASTNode* right;
+	void resolve(Runtime* r) {
+		resolver(r, this);
+	}
+};
