+++ src/pdc.h
@@ -1,0 +1,10 @@
+#define ADD 		'+'
+#define SUBSTRACT	'-'
+#define MULTIPLY	'*'
+#define DIVIDE		'/'
+
+//TODO: Dynamic typing, for now every value is int
+#define PseudoValue int // To enable dynamic typing we need to use one type for every value
+
+#define Program std::vector<ASTNode*>
+
